requirements.txt

flask
pandas
matplotlib
python-dotenv
google-generativeai
